:hammer: Refactor data files path
