--- app/src/main/java/com/rummenigged/popcorntime/data/SeriesRemoteDataSource.kt
@@ -1,5 +1,0 @@
-package com.rummenigged.popcorntime.data
-
-interface SeriesRemoteDataSource {
-   suspend fun fetchSeriesList(page: Int?): List<SeriesRaw>
-}
