+++ app/src/main/java/com/rummenigged/popcorntime/data/common/Raw.kt
@@ -1,4 +1,4 @@
-package com.rummenigged.popcorntime.data
+package com.rummenigged.popcorntime.data.common
 
 interface Raw<SafeType> {
     fun asSafe(): SafeType
