+++ app/src/main/java/com/rummenigged/popcorntime/data/network/api/SeriesApi.kt
@@ -1,5 +1,6 @@
-package com.rummenigged.popcorntime.data
+package com.rummenigged.popcorntime.data.network.api
 
+import com.rummenigged.popcorntime.data.model.SeriesRaw
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
