+++ app/src/main/java/com/rummenigged/popcorntime/data/remoteDataSource/SeriesRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.rummenigged.popcorntime.data.remoteDataSource
+
+import com.rummenigged.popcorntime.data.model.SeriesRaw
+
+interface SeriesRemoteDataSource {
+   suspend fun fetchSeriesList(page: Int?): List<SeriesRaw>
+}
